requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
aiomysql
python-dotenv
python-jose
bcrypt
passlib[bcrypt]
pydantic[email]